backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-stripe==2.7.3
Django==3.2.19
django-4-jet @ https://github.com/Barukimang/django-jet/archive/dev.zip#sha256=73bfdc1ee258ae78c2a1149bea889de8d8a70b1b5807c769c5277f1b40bab499
django-debug-toolbar==3.8.1
django-environ==0.10.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
drf-yasg==1.21.5
feedparser==6.0.10
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
MarkupSafe==2.1.2
packaging==23.1
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.29
ruamel.yaml.clib==0.2.7
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
stripe==4.2.0
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
